+++ deno.json
@@ -10,11 +10,11 @@
     "./extensions/ext-auth-jwt",
     "./extensions/ext-bundler-esbuild",
     "./extensions/ext-cache-redis",
+    "./extensions/ext-css-tailwind",
     "./extensions/ext-mdx",
     "./extensions/ext-node-compat",
     "./extensions/ext-opentelemetry",
-    "./extensions/ext-parser-babel",
-    "./extensions/ext-tailwind"
+    "./extensions/ext-parser-babel"
   ],
   "exclude": [
     "npm/",
@@ -73,6 +73,7 @@
     "./extensions/auth": "./src/extensions/auth/index.ts",
     "./extensions/bundler": "./src/extensions/bundler/index.ts",
     "./extensions/cache": "./src/extensions/cache/index.ts",
+    "./extensions/css": "./src/extensions/css/index.ts",
     "./extensions/parser": "./src/extensions/parser/index.ts",
     "./testing": "./src/testing/index.ts",
     "./testing/assert": "./src/testing/assert.ts",
@@ -158,6 +159,7 @@
     "veryfront/testing/bdd": "./src/testing/bdd.ts",
     "veryfront/extensions/bundler": "./src/extensions/bundler/index.ts",
     "veryfront/extensions/contracts": "./src/extensions/contracts.ts",
+    "veryfront/extensions/css": "./src/extensions/css/index.ts",
     "veryfront/extensions/interfaces": "./src/extensions/interfaces/index.ts",
     "veryfront/extensions/parser": "./src/extensions/parser/index.ts",
     "jose": "npm:jose@5.9.6",
