+++ extensions/ext-css-tailwind/deno.json
@@ -1,5 +1,5 @@
 {
-  "name": "@veryfront/ext-tailwind",
+  "name": "@veryfront/ext-css-tailwind",
   "version": "0.1.0",
   "exports": "./src/index.ts",
   "veryfront": {
@@ -17,7 +17,7 @@
     "@std/assert": "jsr:@std/assert@1",
     "@std/testing/bdd": "jsr:@std/testing@1/bdd",
     "veryfront/extensions": "../../src/extensions/index.ts",
-    "veryfront/extensions/interfaces": "../../src/extensions/interfaces/index.ts"
+    "veryfront/extensions/css": "../../src/extensions/css/index.ts"
   },
   "tasks": {
     "test": "deno test --no-check --allow-all src/"
