+++ src/extensions/css/index.ts
@@ -1,0 +1,13 @@
+/**
+ * CSS category barrel — CSS processor and compiler contracts.
+ *
+ * @module extensions/css
+ */
+
+export type {
+  CSSCompileOptions,
+  CSSCompiler,
+  CSSModuleSource,
+  CSSProcessor,
+  CSSStylesheetSource,
+} from "./css-processor.ts";
